# Add @pattern_f speedup code