Tested on iPhone 12 pro iOS 14.3 

Speed improvement is greatly noted 


Signed-off-by: Bryan Smith <bryan.smith@mobomo.com>