@st0ic-7 2021-03-08
From my tests, stage 3 takes on average 16 seconds