@spacepilotAV 2021-03-08
I have a 6s but haven't had the chance to test this on my device, how fast is it on the 6s (if you have one and don't mind) compared to the regular ~3 minutes on the 6s?