@sbtoonz 2021-03-08
Direct excerpt from @pattern_F "makes the exploit faster (iPhone 12: 65s -> 10s, iPhone 6s: 188s -> 68s)"