@sbtoonz 2021-04-21
It’s been open for like a month with mo attention from the devs so figured
they don’t care

On Wed, Apr 21, 2021 at 4:39 PM ra1n ***@***.***> wrote:

> @sbtoonz <https://github.com/sbtoonz> why did you close this
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/ModernPwner/cicuta_virosa/pull/15#issuecomment-824374374>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AQGFAZ3URAJUFZE46ZA3BFLTJ5AZVANCNFSM4YWCRCIA>
> .
>